+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/dialog/DefaultActionDialogBuilder.java
@@ -23,7 +23,6 @@ import is.codion.swing.common.ui.component.panel.BorderLayoutPanelBuilder;
 import is.codion.swing.common.ui.component.panel.PanelBuilder;
 import is.codion.swing.common.ui.control.Controls;
 import is.codion.swing.common.ui.key.KeyEvents;
-import is.codion.swing.common.ui.layout.Layouts;
 
 import org.jspecify.annotations.Nullable;
 
@@ -184,7 +183,7 @@ class DefaultActionDialogBuilder<B extends ActionDialogBuilder<B>> extends Abstr
 
 	private JPanel createPanel(JPanel buttonPanel) {
 		JPanel buttonBasePanel = PanelBuilder.builder()
-						.layout(Layouts.flowLayout(buttonPanelConstraints))
+						.flowLayout(buttonPanelConstraints)
 						.add(buttonPanel)
 						.border(buttonPanelBorder)
 						.build();